+++ Runtime/Extensions/Tools.cs
@@ -5,13 +5,8 @@ using System.Linq;
 
 namespace Mane.Extensions
 {
-    public static class BaseExtensions
+    public static class Tools
     {
-        public static Version ToVersion(this string v)
-        {
-            return new Version(v);
-        }
-
         public static IEnumerable<T> GetEnumValues<T>()
         {
             return Enum.GetValues(typeof(T)).Cast<T>();